Models Manual and Proportional Control. Heat is supplied through the PWM and there is heat loss to a room at constant temperature.
There are two versions. In one, the measured temperature and the TEC temperature are the same. This corresponds to heating a small, very responsive mass. In the other, the thermometer has a mass and takes some time to heat up to the temperature of the TEC. This leads to phase lag and oscillation.