TEC code for Angstrom experiment.
Angstrom experiment to sinusoidally vary temperature and measure the temperature at different locations along a cylindrical rod. This code has 4 thermistors. One used for PID control of the TEC and 3 to measure temperature the temperature along the rod. The arduino code is "SerialTestNov1a.ino" and the Matlab code, which is a GUI, is TEC_V7.mlapp.

The time and the temperature from the 4 thermistors are stored in a file. The parameters controlling the temperature are recorded each time a parameter, such as the set temperature, Kp, KI, sin amplitude or sin period are changed.